requirements.txt
nomadicml>=0.1.40
python-dotenv>=1.0.0
boto3>=1.34.0
